+++ src/main/java/dev/louis/nebula/api/mana/storage/ManaStorage.java
@@ -1,6 +1,6 @@
 package dev.louis.nebula.api.mana.storage;
 
 public interface ManaStorage {
-    float getMana();
-    float getCapacity();
+    long getMana();
+    long getCapacity();
 }
